backend/requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
python-multipart
weasyprint

# Security and Authentication
passlib[argon2]
python-jose[cryptography]

# Database - PostgreSQL
SQLAlchemy
psycopg2-binary

# Testing
pytest
httpx
